+++ gdsfactory/simulation/gmeep/write_sparameters_meep_batch.py
@@ -29,7 +29,7 @@ temp_dir_default = Path(sparameters_path) / "temp"
 
 
 @pydantic.validate_arguments
-def write_sparameters_meep_mpi_pool(
+def write_sparameters_meep_batch(
     jobs: List[Dict],
     cores_per_run: int = 2,
     total_cores: int = 4,
@@ -39,9 +39,9 @@ def write_sparameters_meep_mpi_pool(
     layer_stack: LayerStack = LAYER_STACK,
     **kwargs,
 ) -> List[Path]:
-    """Write Sparameters and returns the filepaths.
+    """Write Sparameters for a batch of jobs using MPI and returns results filepaths.
     Given a list of write_sparameters_meep keyword arguments (the "jobs"),
-    launches them in different cores
+    launches them in different cores using MPI
     where each simulation runs with "cores_per_run" cores
     If there are more simulations than cores each batch runs sequentially
 
@@ -57,9 +57,27 @@ def write_sparameters_meep_mpi_pool(
         layer_stack:
 
     keyword Args:
-        overwrite: overwrites stored simulation results.
+        resolution: in pixels/um (30: for coarse, 100: for fine)
+        port_symmetries: Dict to specify port symmetries, to save number of simulations
+        dirpath: directory to store Sparameters
+        layer_stack: LayerStack class
+        port_margin: margin on each side of the port
+        port_monitor_offset: offset between monitor GDS port and monitor MEEP port
+        port_source_offset: offset between source GDS port and source MEEP port
+        filepath: to store pandas Dataframe with Sparameters in CSV format.
+        animate: saves a MP4 images of the simulation for inspection, and also
+            outputs during computation. The name of the file is the source index
+        lazy_parallelism: toggles the flag "meep.divide_parallel_processes" to
+            perform the simulations with different sources in parallel
         dispersive: use dispersive models for materials (requires higher resolution)
+        xmargin: left and right distance from component to PML.
+        xmargin_left: west distance from component to PML.
+        xmargin_right: east distance from component to PML.
+        ymargin: top and bottom distance from component to PML.
+        ymargin_top: north distance from component to PML.
+        ymargin_bot: south distance from component to PML.
         extend_ports_length: to extend ports beyond the PML
+        layer_stack: Dict of layer number (int, int) to thickness (um)
         zmargin_top: thickness for cladding above core
         zmargin_bot: thickness for cladding below core
         tpml: PML thickness (um)
@@ -168,12 +186,12 @@ def write_sparameters_meep_mpi_pool(
     return filepaths
 
 
-write_sparameters_meep_mpi_pool_lr = gf.partial(
-    write_sparameters_meep_mpi_pool, ymargin_top=3, ymargin_bot=3
+write_sparameters_meep_batch_lr = gf.partial(
+    write_sparameters_meep_batch, ymargin_top=3, ymargin_bot=3
 )
 
-write_sparameters_meep_mpi_pool_lt = gf.partial(
-    write_sparameters_meep_mpi_pool, ymargin_bot=3, xmargin_right=3
+write_sparameters_meep_batch_lt = gf.partial(
+    write_sparameters_meep_batch, ymargin_bot=3, xmargin_right=3
 )
 
 
@@ -189,7 +207,7 @@ if __name__ == "__main__":
         for i in range(1, 4)
     ]
 
-    filepaths = write_sparameters_meep_mpi_pool(
+    filepaths = write_sparameters_meep_batch(
         jobs=jobs,
         cores_per_run=4,
         total_cores=8,